requirements.txt
autopep8==1.6.0
mysql-connector-python==8.0.27
protobuf==3.19.0
pycodestyle==2.8.0
toml==0.10.2
